Add an audio element to topics
There's still some styling to be done, and probably some configuration,
but the main functionality is in place.

It turns out that even though the <audio> element is supported by all
modern browsers, they can't quite agree on which audio formats it should
handle. Consequently, I've included the SoundManager 2 library, which
provides both HTML5 and Flash playback with a JavaScript API.